% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DetAbu}
\alias{DetAbu}
\title{Estimating detected species relative abundance}
\usage{
DetAbu(x, zero)
}
\arguments{
\item{x}{a vector of species abundance frequency}

\item{zero}{reserve zero frequency or not. Default is \code{FALSE}.}
}
\value{
a numerical vector
}
\description{
Estimating detected species relative abundance
}

