% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DetInc}
\alias{DetInc}
\title{Estimating detected species incidence probability}
\usage{
DetInc(y, zero)
}
\arguments{
\item{y}{a vector of species incidence frequency}

\item{zero}{reserve zero frequency or not. Default is \code{FALSE}.}
}
\value{
a numerical vector
}
\description{
\code{DetInc} Estimating detected species incidence probability
}

