\name{JADE-package}
\alias{JADE-package}
\docType{package}
\title{
JADE and other BSS methods as well as some BSS performance criteria
}
\description{
The package ports JF Cardoso's JADE algorithm as well as his function for joint diagonalization. The package also includes several other blind source separation (BSS) methods, like AMUSE and SOBI, as well as some criteria for performance evaluation of BSS algorithms. 
}
\details{
\tabular{ll}{
Package: \tab JADE\cr
Type: \tab Package\cr
Version: \tab 1.9-9\cr
Date: \tab 2013-10-10\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Klaus Nordhausen, Jean-Francois Cardoso, Jari Miettinen, Hannu Oja, Esa Ollila, Sara Taskinen

Maintainer: Klaus Nordhausen <klaus.nordhausen@utu.fi>

}
\keyword{package}
