\name{print.bss}
\alias{print.bss}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Printing an Object of Class bss}
\description{
Prints an object of class bss. It prints all elements of the list of class bss except the component S
which is the source matrix.
}
\usage{
\method{print}{bss}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{object of class bss.}
  \item{\dots}{further arguments to be passed to or from methods.}
}
\author{Klaus Nordhausen}
\keyword{methods}
