% Generated by roxygen2 (4.0.1): do not edit by hand
\name{UndInc}
\alias{UndInc}
\title{Estimating undetected species incidence probability}
\usage{
UndInc(y)
}
\arguments{
\item{y}{a vector of species incidence frequency.}
}
\value{
a numerical vector
}
\description{
\code{UndInc} Estimating undetected species incidence probability
}

