\name{JADE-package}
\alias{JADE-package}
\docType{package}
\title{
JADE and other BSS methods as well as some BSS performance criteria
}
\description{
The package ports JF Cardoso's JADE algorithm as well as his function for joint diagonalization. 
There are also several other blind source separation (BSS) methods like AMUSE and SOBI  as well as some criteria for performance evaluation of BSS algorithms. 
}
\details{
\tabular{ll}{
Package: \tab JADE\cr
Type: \tab Package\cr
Version: \tab 1.1-0\cr
Date: \tab 2012-02-14\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Klaus Nordhausen, Jean-Francois Cardoso, Jari Miettinen, Hannu Oja, Esa Ollila, Sara Taskinen

Maintainer: Klaus Nordhausen <klaus.nordhausen@uta.fi>

}
\keyword{package}
