\name{JADE-package}
\alias{JADE-package}
\docType{package}
\title{
Blind Source Separation Methods Based on Joint Diagonalization and Some BSS Performance Criteria
}
\description{
Cardoso's JADE algorithm as well as his functions for joint diagonalization are ported to R. Also several other blind source separation (BSS) methods, like AMUSE and SOBI, and some criteria for performance evaluation of BSS algorithms, are given.
}
\details{
\tabular{ll}{
Package: \tab JADE\cr
Type: \tab Package\cr
Version: \tab 1.9-93\cr
Date: \tab 2015-08-14\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Klaus Nordhausen, Jean-Francois Cardoso, Jari Miettinen, Hannu Oja, Esa Ollila, Sara Taskinen

Maintainer: Klaus Nordhausen <klaus.nordhausen@utu.fi>

}
\keyword{package}
