\name{JADE-package}
\alias{JADE-package}
\docType{package}
\title{
\packageTitle{JADE}
}
\description{
\packageDescription{JADE}
}
\details{
\packageDESCRIPTION{JADE}

\packageIndices{JADE}
}

\references{
\cite{Miettinen, J., Nordhausen, K. and Taskinen, S. (2017), Blind Source Separation Based on Joint Diagonalization in R: The Packages JADE and BSSasymp, \emph{Journal of Statistical Software}, \bold{76}, 1--31, <doi:10.18637/jss.v076.i02>.}
}

\author{
\packageAuthor{JADE}

Maintainer: \packageMaintainer{JADE}

}
\keyword{package}
