\name{JADE-package}
\alias{JADE-package}
\docType{package}
\title{
Blind Source Separation Methods Based on Joint Diagonalization and Some BSS Performance Criteria
}
\description{
Cardoso's JADE algorithm as well as his functions for joint diagonalization are ported to R. Also several other blind source separation (BSS) methods, like AMUSE and SOBI, and some criteria for performance evaluation of BSS algorithms, are given.
}
\details{
\tabular{ll}{
Package: \tab JADE\cr
Type: \tab Package\cr
Version: \tab 2.0-0\cr
Date: \tab 2017-01-09\cr
License: \tab GPL (>= 2)\cr
}

}

\references{
\cite{Miettinen, J., Nordhausen, K. and Taskinen, S. (2017), Blind Source Separation Based on Joint Diagonalization in R: The Packages JADE and BSSasymp, \emph{Journal of Statistical Software}, \bold{76}, 1--31, <doi:10.18637/jss.v076.i02>.}
}

\author{
Klaus Nordhausen, Jean-Francois Cardoso, Jari Miettinen, Hannu Oja, Esa Ollila, Sara Taskinen

Maintainer: Klaus Nordhausen <klaus.nordhausen@utu.fi>

}
\keyword{package}
