\name{JADE-package}
\alias{JADE-package}
\docType{package}
\title{
JADE and ICA performance criteria
}
\description{
The package ports J. F. Cardoso's ICA algorithm JADE to R as well as his function for joint diagonalization. Furthermore provides the package
several criteria to evaluate ICA algorithms in simulation studies. 
}
\details{
\tabular{ll}{
Package: \tab JADE\cr
Type: \tab Package\cr
Version: \tab 1.0-2\cr
Date: \tab 2009-04-20\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Klaus Nordhausen, Jean-Francois Cardoso, Hannu Oja, Esa Ollila

Maintainer: Klaus Nordhausen <klaus.nordhausen@uta.fi>

}
\keyword{package}
